mir 23 - 1.28

+++ lib/Utils/TimUtil.dart
@@ -15,15 +15,15 @@ class TimUtil {
       time = format.format(date);
     } else if (diff.inDays > 0 && diff.inDays < 7) {
       if (diff.inDays == 1) {
-        time = '${diff.inDays} day';
+        time = diff.inDays.toString() + ' day';
       } else {
-        time = '${diff.inDays} days';
+        time = diff.inDays.toString() + ' days';
       }
     } else {
       if (diff.inDays == 7) {
-        time = '${(diff.inDays / 7).floor()} wk';
+        time = (diff.inDays / 7).floor().toString() + ' wk';
       } else {
-        time = '${(diff.inDays / 7).floor()} weeks';
+        time = (diff.inDays / 7).floor().toString() + ' weeks';
       }
     }
 
@@ -31,7 +31,7 @@ class TimUtil {
   }
 
   static int currentTimeInSeconds() {
-    var ms = (DateTime.now()).millisecondsSinceEpoch;
+    var ms = (new DateTime.now()).millisecondsSinceEpoch;
     return (ms / 1000).round();
   }
 
@@ -74,7 +74,7 @@ class TimUtil {
     DateTime date = DateTime.fromMillisecondsSinceEpoch(timestamp * 1000);
     final date2 = DateTime.now();
     final difference = date2.difference(date);
-    print("diff in minutes ${difference.inMinutes}");
+    print("diff in minutes " + difference.inMinutes.toString());
     if (difference.inMinutes > 10) {
       return true;
     } else {
@@ -94,16 +94,15 @@ class TimUtil {
     String seconds;
     final minutes = duration.inMinutes.remainder(60);
     final sec = duration.inSeconds.remainder(60);
-    if (sec < 10) {
+    if (sec < 10)
       seconds = '0$sec';
-    } else {
+    else
       seconds = '$sec';
-    }
     return '$minutes:$seconds';
   }
 
   static String stringForSeconds(double seconds) {
-    if (seconds == 0) return "00:00";
+    if (seconds == null || seconds == 0) return "00:00";
     return '${(seconds ~/ 60)}:${(seconds.truncate() % 60).toString().padLeft(2, '0')}';
   }
 
@@ -116,7 +115,9 @@ class TimUtil {
       sec = sec % 60;
     }
 
-    return "${min >= 10 ? min.toString() : '0$min'}:${sec >= 10 ? sec.toString() : '0$sec'}";
+    return (min >= 10 ? min.toString() : '0' + min.toString()) +
+        ":" +
+        (sec >= 10 ? sec.toString() : '0' + sec.toString());
   }
 
   static int parseDuration(String s) {
@@ -133,7 +134,7 @@ class TimUtil {
     micros = (double.parse(parts[parts.length - 1]) * 1000000).round();
     int tim = Duration(hours: hours, minutes: minutes, microseconds: micros)
         .inSeconds;
-    print("current video duration = $tim");
+    print("current video duration = " + tim.toString());
     return tim;
   }
 }
